Fix the caching logic

+++ src/oss-characteristics/CharacteristicTool.cs
@@ -112,6 +112,7 @@ namespace Microsoft.CST.OpenSource
 
             var downloadTool = new DownloadTool();
             // ensure that the cache directory has the required package, download it otherwise
+            bool cached = !string.IsNullOrEmpty(targetDirectoryName);
             if (!string.IsNullOrEmpty(targetDirectoryName))
             {
                 directoryNames.AddRange(await downloadTool.EnsureDownloadExists(purl, targetDirectoryName));
@@ -125,7 +126,10 @@ namespace Microsoft.CST.OpenSource
                     Logger.Trace("Removing directory {0}", directoryName);
                     try
                     {
-                        Directory.Delete(directoryName, true);
+                        if (!cached)
+                        {
+                            Directory.Delete(directoryName, true);
+                        }
                     }
                     catch (Exception ex)
                     {
